+++ app/sitemap.ts
@@ -82,6 +82,12 @@ export default function sitemap(): MetadataRoute.Sitemap {
       priority: 0.8,
     },
     {
+      url: absoluteUrl("/guides/how-many-bricks-do-i-need"),
+      lastModified: new Date("2026-08-13"),
+      changeFrequency: "monthly",
+      priority: 0.8,
+    },
+    {
       url: absoluteUrl("/guides/how-much-gravel-do-i-need"),
       lastModified: new Date("2026-08-13"),
       changeFrequency: "monthly",